requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
astroid==2.4.0
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.2
backcall==0.1.0
bcolz==1.2.1
bleach==3.1.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
flann==1.6.13
future==0.18.2
gast==0.3.3
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
ijson==3.0.4
imagecodecs==2020.2.18
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.5
notebook==6.0.3
num2words==0.5.10
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pooch==1.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.5.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.1
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.90
six==1.14.0
tensorboard==1.14.0
tensorboardX==2.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.5.7
toml==0.10.0
torch @ file:///home/adnen/Downloads/torch-1.3.1%2Bcu100-cp36-cp36m-linux_x86_64.whl
torchfile==0.1.0
torchtext==0.6.0
torchvision @ file:///home/adnen/Downloads/torchvision-0.4.2%2Bcu100-cp36-cp36m-linux_x86_64.whl
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
